requirements.txt
pywebview
flask-cors
wmi
pypdf
escpos
psutil
werkzeug
python-barcode
flask
django-qrcode
requests
pillow
customtkinter
